issue #101 split base classes and exceptions from Fortran2003.py into utils.py
