+++ src/fparser/two/tests/fortran2008/test_end_submodule_stmt_r1119.py
@@ -38,7 +38,7 @@
 
 '''
 import pytest
-from fparser.two.Fortran2003 import NoMatchError
+from fparser.two.utils import NoMatchError
 from fparser.two.Fortran2008 import End_Submodule_Stmt
