+++ src/fparser/two/tests/fortran2008/test_parent_identifier_r1118.py
@@ -45,7 +45,7 @@
 
 '''
 import pytest
-from fparser.two.Fortran2003 import NoMatchError
+from fparser.two.utils import NoMatchError
 from fparser.two.Fortran2008 import Parent_Identifier
