+++ src/fparser/two/tests/fortran2008/test_program_unit_r202.py
@@ -38,7 +38,7 @@ submodules in Fortran2008 for the Program-unit rule.
 '''
 
 import pytest
-from fparser.two.Fortran2003 import NoMatchError
+from fparser.two.utils import NoMatchError
 from fparser.api import get_reader
 from fparser.two.Fortran2008 import Program_Unit
